https://jira.csiro.au/browse/ASKAPSDP-1680 Changes to match the new VisDatagramADE.h (channel is now 1-based and frequency is now double)

+++ Code/Components/Services/correlatorsim/current/simplayback/FreqIndex.h
@@ -54,7 +54,7 @@ class FreqIndex
 
         uint32_t channel;
 
-        float freq;
+        double freq;
 };
 
 };
